more refactoring, created a special directory for models
