+++ lib/models/score.dart
@@ -1,0 +1,7 @@
+class Score {
+  final String category;
+  final int value;
+  final DateTime date;
+
+  Score({required this.category, required this.value, required this.date});
+}
